JogadorID e CenaID movidos de espaço de nomes para Identificadores
