+++ duendesproj/Assets/scripts/Identificadores.cs
@@ -1,0 +1,14 @@
+namespace Identificadores
+{
+    public enum JogadorID {J1, J2, J3, J4};
+
+    public enum CenaID
+    {
+        Tabuleiro           = 0,
+        QuebraBotao         = 1,
+        PescaEscorrega      = 2,
+        TelaCarregamento    = 3,
+        MenuInicial         = 4,
+        // TODO: Colocar outros aqui
+    };
+}
