+++ duendesproj/Assets/scripts/Identificadores.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e35fc80911206f74911fdb87030dd85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
