+++ duendesproj/Assets/scripts/gerenciadores/utilidades_dsv/FuncoesGGTestes.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Identificadores;
 
 namespace Gerenciadores.DsvUtils
 {
